+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/NumberSchemaSpec.scala
@@ -35,7 +35,7 @@ class NumberSchemaSpec extends UnitSpec {
 
   it should "expand by incrementing values" in {
     MaxNumValueProperty(Some(11))
-      .expandTo(MaxNumValueProperty((Some(12))))
+      .expandTo(Some(MaxNumValueProperty((Some(12)))))
       .maxNumValue shouldBe Some(12)
   }
 
@@ -53,7 +53,7 @@ class NumberSchemaSpec extends UnitSpec {
 
   it should "expand by decrementing values" in {
     MinNumValueProperty(Some(11))
-      .expandTo(MinNumValueProperty((Some(10))))
+      .expandTo(Some(MinNumValueProperty((Some(10)))))
       .minNumValue shouldBe Some(10)
   }
 
@@ -135,19 +135,19 @@ class NumberSchemaSpec extends UnitSpec {
 
   it should "expand by division by 2s" in {
     NumMultipleOfProperty(Some(4.0))
-      .expandTo(NumMultipleOfProperty((Some(2.0))))
+      .expandTo(Some(NumMultipleOfProperty((Some(2.0)))))
       .multiple shouldBe Some(2.0)
   }
 
   it should "stop expanding after MaxExpandRounds" in {
     NumMultipleOfProperty(Some(1048576.0))
-      .expandTo(NumMultipleOfProperty((Some(2.0))))
+      .expandTo(Some(NumMultipleOfProperty((Some(2.0)))))
       .multiple shouldBe None
   }
 
   it should "not expand if already covered" in {
     NumMultipleOfProperty(Some(3.5))
-      .expandTo(NumMultipleOfProperty((Some(7.0))))
+      .expandTo(Some(NumMultipleOfProperty((Some(7.0)))))
       .multiple shouldBe Some(3.5)
   }
 
@@ -293,6 +293,8 @@ class NumberSchemaSpec extends UnitSpec {
 
   it should "expand to be compatible with a similar schema" in {
     val schema = NumberSchema(2)
-    NumberSchema(1).expandTo(schema).isCompatibleWith(schema) shouldBe true
+    NumberSchema(1)
+      .expandTo(Some(schema))
+      .isCompatibleWith(schema) shouldBe true
   }
 }