+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/ObjectSchemaSpec.scala
@@ -189,12 +189,12 @@ class ObjectSchemaSpec extends UnitSpec {
   it should "expand to add new keys" in {
     val schema =
       ObjectSchema(Map("foo" -> BooleanSchema(), "bar" -> BooleanSchema()))
-    objectSchema.expandTo(schema).isCompatibleWith(schema) shouldBe true
+    objectSchema.expandTo(Some(schema)).isCompatibleWith(schema) shouldBe true
   }
 
   it should "expand to combine types in the same key" in {
     val schema1 = ObjectSchema(Map("foo" -> IntegerSchema(1)))
     val schema2 = ObjectSchema(Map("foo" -> IntegerSchema(2)))
-    schema1.expandTo(schema2).isCompatibleWith(schema2) shouldBe true
+    schema1.expandTo(Some(schema2)).isCompatibleWith(schema2) shouldBe true
   }
 }