+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add a property to track the percentage of true Boolean values
+- Allow for oblivious expansion without another schema
 
 ### Changed
 - Allow reset of min/max length for strings with format